pyproject.toml
[tool.poetry]
name = "challenge-python-v1"
version = "0.1.0"
description = "Python application for Dinosaurs Aficionados which is going to be used to maintain and provide various information about all kinds of Dinosaurs."
authors = ["Katarina Kovacova <katarinakovacova100@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
psycopg2 = "^2.9.9"
pytest = "^8.3.2"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
